Moved Storage trait into storage module root, where it belongs.
