+++ src/storage/mod.rs
@@ -1,7 +1,13 @@
 use async_trait::async_trait;
-use futures_core::Stream;
+use bytes::Buf;
+use futures_core::{Future, Stream};
+use futures_lite::AsyncWrite;
 use num::{cast::FromPrimitive, CheckedSub, ToPrimitive, Unsigned};
-use std::{cmp, iter::Sum, ops::RangeBounds};
+use std::{
+    cmp,
+    iter::Sum,
+    ops::{Deref, RangeBounds},
+};
 
 pub mod commit_log;
 
@@ -97,10 +103,42 @@ pub trait AsyncConsume {
     async fn close(self) -> Result<(), Self::ConsumeError>;
 }
 
-pub trait SizedStorage {
+pub trait Sizable {
     type Size: Unsigned + FromPrimitive + Sum + Ord;
 
     fn size(&self) -> Self::Size;
+}
+
+#[async_trait(?Send)]
+pub trait Storage:
+    AsyncTruncate<Mark = Self::Position, TruncError = Self::Error>
+    + AsyncConsume<ConsumeError = Self::Error>
+    + Sizable
+{
+    type Content: Deref<Target = [u8]>;
+
+    type Write: AsyncWrite + Unpin;
+
+    type Position;
+
+    type Error: std::error::Error;
+
+    async fn read(
+        &mut self,
+        position: &Self::Position,
+        size: &Self::Size,
+    ) -> Result<Self::Content, Self::Error>;
+
+    async fn append<B, S, W, F, T>(
+        &mut self,
+        byte_stream: &mut S,
+        write_fn: &mut W,
+    ) -> Result<(Self::Position, T), Self::Error>
+    where
+        B: Buf,
+        S: Stream<Item = B> + Unpin,
+        W: FnMut(&mut S, &mut Self::Write) -> F,
+        F: Future<Output = std::io::Result<T>>;
 }
 
 pub mod impls;
